AsyncElement: Move api::async_elment to a module
This commit moves AsyncElement,  AsyncElementLink and its related
tests into a new file, async_element.rs, that lives under the api
directory.
